requirements.txt
-i https://pypi.org/simple
autopep8==2.0.2
black==23.3.0
certifi==2023.5.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
flake8==6.0.0
idna==3.4 ; python_version >= '3.5'
isort==5.12.0
mccabe==0.7.0 ; python_version >= '3.6'
mypy-extensions==1.0.0 ; python_version >= '3.5'
packaging==23.1 ; python_version >= '3.7'
pathspec==0.11.1 ; python_version >= '3.7'
platformdirs==3.8.0 ; python_version >= '3.7'
pycodestyle==2.10.0 ; python_version >= '3.6'
pyflakes==3.0.1 ; python_version >= '3.6'
pyyaml==6.0
requests==2.31.0
tomli==2.0.1 ; python_version < '3.11'
urllib3==2.0.3 ; python_version >= '3.7'
